+++ plugins/wiggle/src/shared/YScaleBar.tsx
@@ -3,11 +3,13 @@ import { useTheme } from '@mui/material'
 import { observer } from 'mobx-react'
 import { axisPropsFromTickScale, Axis, LEFT, RIGHT } from 'react-d3-axis-mod'
 
-export default observer(function ({
+type Ticks = ReturnType<typeof axisPropsFromTickScale>
+
+export default observer(function YScaleBar({
   model,
   orientation,
 }: {
-  model: { ticks?: ReturnType<typeof axisPropsFromTickScale> }
+  model: { ticks?: Ticks }
   orientation?: string
 }) {
   const { ticks } = model
